Reorganizing the structure of the owner directory such that the different components have their own directories.
